+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
  - in instant mode (no exec time) the function activate from pin,i2c,spi and async now blocks and waits till the uC activated the interface 
+ - api will no longer check if an interface is active before writing to it (creates unwanted corner cases) => the uC is responcible for reporting that error!
 
 ### Removed